src/main/java/java8/com/insightfullogic/java8/answers/chapter3/FilterUsingReduce.java
package java8.com.insightfullogic.java8.answers.chapter3;

import java.util.ArrayList;
import java.util.List;
import java.util.function.Predicate;
import java.util.stream.Stream;

/**
 * Advanced Exercises Question 1
 */
public class FilterUsingReduce {

    public static <I> List<I> filter(Stream<I> stream, Predicate<I> predicate) {
        List<I> initial = new ArrayList<>();
        List<I> resultList = new ArrayList<>();

        return stream.reduce(initial,
             (List<I> acc, I element) -> {
                 if (predicate.test(element)) {
                    resultList.add(element);
                    return resultList;
                }
                return acc;
             },
             FilterUsingReduce::combineLists);
    }

    private static <I> List<I> combineLists(List<I> left, List<I> right) {
    	// We are copying left to new list to avoid mutating it. 
    	List<I> newLeft = new ArrayList<>(left);
    	newLeft.addAll(right);
        return newLeft;
    }

}
